Updates and fixes

+++ dev-java/sun-jdk/sun-jdk-1.4.2.01.ebuild
@@ -1,11 +1,6 @@
 # Copyright 1999-2003 Gentoo Technologies, Inc.
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/dev-java/sun-jdk/sun-jdk-1.4.2.01.ebuild,v 1.1 2003/08/30 18:52:50 prez Exp $
-
-# Since This Ebuild Has FETCH restrictions:
-# You need to download this file from 
-# http://java.sun.com/j2se/1.4.2/download.html 
-# and copy it on your distfiles directory and emerge it again
+# $Header: /var/cvsroot/gentoo-x86/dev-java/sun-jdk/sun-jdk-1.4.2.01.ebuild,v 1.2 2003/09/12 04:40:27 strider Exp $
 
 IUSE="doc gnome kde mozilla"
 
@@ -14,7 +9,7 @@ inherit java nsplugins
 At="j2sdk-1_4_2_01-linux-i586.bin"
 S="${WORKDIR}/j2sdk1.4.2_01"
 DESCRIPTION="Sun's J2SE Development Kit, version 1.4.2"
-HOMEPAGE="http://java.sun.com/j2se/1.4.2/download.html"
+HOMEPAGE="http://java.sun.com/products/archive/j2se/1.4.2/index.html"
 SRC_URI=${At}
 RESTRICT="fetch"
 SLOT="1.4"
@@ -44,8 +39,13 @@ pkg_nofetch() {
 }
 
 src_unpack() {
-	startAt=`grep -aonm 1 ELF  ${DISTDIR}/${At} | cut -d: -f1`
-	tail +${startAt} ${DISTDIR}/${At} > install.sfx
+	if [ ! -r ${DISTDIR}/${At} ]; then
+		eerror "cannot read ${At}. Please check the permission and try again."
+		die
+	fi
+	testExp=`echo -e "\0177\0105\0114\0106\0001\0001\0001"`
+	startAt=`grep -aonm 1 ${testExp}  ${DISTDIR}/${At} | cut -d: -f1`
+	tail -n +${startAt} ${DISTDIR}/${At} > install.sfx
 	chmod +x install.sfx
 	./install.sfx || die
 	rm install.sfx
